Fixed FLANN indices for 32 Bit / Visual Studio

+++ include/lvr/reconstruction/SearchTreeFlann.hpp
@@ -83,7 +83,7 @@ protected:
     size_t															m_numPoints;
     flann::Matrix<float>  	 										m_flannPoints;
 
-    vector<ulong>													m_ind;
+    vector<int>														m_ind;
     vector<float>													m_dst;
 
 }; // SearchTreeFlann
